add range tests
- Replace older maxmin test code with range. This only tests min/max
and not the reading.
- L23 was incorrectly in raw and is now in quantity.
- Misc minor edits.
