+++ src/server/test/web/readingsLineMeterRangeFlow.js
@@ -1,0 +1,32 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+/*
+    This file tests the readings retrieval API line chart flow meters.
+    See: https://github.com/OpenEnergyDashboard/DesignDocs/blob/main/testing/testing.md for information.
+*/
+
+const { chai, mocha, app } = require('../common');
+const Unit = require('../../models/Unit');
+const { prepareTest,
+    parseExpectedCsv,
+    expectReadingToEqualExpected,
+    getUnitId,
+    ETERNITY,
+    METER_ID } = require('../../util/readingsUtils');
+
+mocha.describe('readings API', () => {
+    mocha.describe('readings test, test if data returned by API is as expected', () => {
+        mocha.describe('for line charts', () => {
+            mocha.describe('for flow meters', () => {
+
+                // Add LR8 here
+
+                // Add LR25 here
+
+                // Add LR26 here
+            });
+        });
+    });
+});
