+++ src/server/test/web/readingsLineMeterRangeRaw.js
@@ -1,0 +1,44 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+/*
+    This file tests the readings retrieval API for line chart raw meters.
+    See: https://github.com/OpenEnergyDashboard/DesignDocs/blob/main/testing/testing.md for information.
+*/
+
+const { chai, mocha, app } = require('../common');
+const Unit = require('../../models/Unit');
+const { prepareTest,
+    parseExpectedCsv,
+    expectReadingToEqualExpected,
+    createTimeString,
+    getUnitId,
+    ETERNITY,
+    METER_ID,
+    unitDatakWh,
+    conversionDatakWh,
+    meterDatakWh } = require('../../util/readingsUtils');
+
+mocha.describe('readings API', () => {
+    mocha.describe('readings test, test if data returned by API is as expected', () => {
+        mocha.describe('for line charts', () => {
+            mocha.describe('for range (min/max)', () => {
+                mocha.describe('for raw meters', () => {
+
+                    // Add LR9 here
+
+                    // Add LR14 here
+
+                    // Add LR15 here
+
+                    // Add LR16 here
+
+                    // Add LR17 here
+
+                    // Add LR22 here
+                });
+            });
+        });
+    });
+});
